+++ src/pages/campaigns/hooks/useCampaignMuations.jsx
@@ -1,9 +1,11 @@
-import { editCampaign, saveCampaign } from "@/domains/campaign/api/campaign";
-import { useMutation } from "@tanstack/react-query";
+import { deleteCampaign, editCampaign, likeCampaign, saveCampaign } from "@/domains/campaign/api/campaign";
+import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { useNavigate } from "react-router-dom";
 
 export function useCampaignMutations() {
   const navigate = useNavigate();
+  const queryClient = useQueryClient();
+
   const updateCampaignMutation = useMutation({
     mutationFn: saveCampaign,
     onSuccess: ({ success }) => {
@@ -22,5 +24,25 @@ export function useCampaignMutations() {
       }
     },
   });
-  return { updateCampaignMutation, editCampaignMutation };
+
+  const likeCampaignMutation = useMutation({
+    mutationFn: likeCampaign,
+    onSuccess: ({ success }) => {
+      if (success) {
+        queryClient.invalidateQueries({ queryKey: ["campaign"] });
+      }
+    },
+  });
+
+  const deleteCampaignMutation = useMutation({
+    mutationFn: deleteCampaign,
+    onSuccess: ({ success }) => {
+      if (success) {
+        queryClient.invalidateQueries({ queryKey: ["campaign"] });
+        navigate(-1);
+      }
+    },
+  });
+
+  return { updateCampaignMutation, editCampaignMutation, likeCampaignMutation, deleteCampaignMutation };
 }
